52 : made list cell a separate class for sanity
